"Fix" bipolar param bug
This bug is on Korg's side -- we're not getting the right messages to
OSC_PARAM, and there's no way to tell the difference between the right
messages and the wrong ones. So in this case, we'll just assume that if
the first message we get is a 0, it's wrong, but this will break presets
that use Release = -100%. Nothing we can do about that other than to
tell users to use Release = -99% instead for the time being, which is
essentially identical.