Upgrade to C++17, CMSIS5; clean up ARM workarounds
I prefer working with newer versions of C++, so upgrade with a compiler
option. This requires a newer version of gcc than Korg prefers, so we've
also got to update that within the Korg repo and update to the new path
in our Makefile, as well as removing the -mstructure-size-boundary
option, which is deprecated.

CMSIS version 4 uses the 'register' storage class specifier, which no
longer exists in C++17. CMSIS5 fixes this issue, so we upgrade to that;
however, CMSIS5 breaks up includes, so we need to fix those include
paths in the Makefile, too.

Finally, we've had dummy __aeabi_atexit function and __dso_handle
objects in our main file for far too long. Turns out we can eliminate
that issue with the -fno-use-cxa-atexit option.